Suppress an incorrect cppcheck warning
Suppress a cppcheck warning where cppcheck doesnt understand that the
function it complains about is a juce function that are intended to use.
cppcheck want to use expm1 (exponantional minus 1) inseted of
exp(argument) -1.0. However the juce function expm1 doesnt exist and the
juce function is supposed to be much faster than the normal exp().